Test END only once in evaluator loop
The evaluator loop had a case for REB_END in its switch, which had
to be turned into an IS_END test prior to the switch.  This pointed out
that in actuality, the only time an END test should happen is at the
very start of the call...and can be short-circuited.

The reason that END testing is not appropriate after the initial check
is because an arbitrary evaluation can resize a block such that the
previous evaluation index is beyond the end.  The actual length must
be looked at to avoid crashes, because the END is only positioned
exactly at the end... so END_FLAG is produced if the index is seen to
be past that point--not because there was an END element.