Adapt PARAM and REFINE for frameless natives
The debug build variant of PARAM and REFINE would stash the
pointers to the values for that parameter or argument, even though
the release build optimized the structures out as constants.  However,
the ability to refer to parameters by index is valuable to frameless
natives as well (e.g. for reporting error messages).  This makes the
debug build back off trying to cache pointers to the actual arguments
in the case of a frameless native.