Remove stack overflow checking from Make_Series
Checking for a C stack overflow is done using a somewhat tenuous and
non-C-standard mechanic, which should be reviewed.  It involves doing a
comparison of pointer numbers manually in certain sporadic spots within
the code.

The main place this should be done is from within recursive functions and
places where the stack can reasonably be thought to overflow.  There was
a call in Make_Series which was presumably there based on the idea that
"it gets called a lot", but that's not the metric of where such checks should
be done.