Exploit END as low bit zero in chunked stack
This is the first "clever" application of the new strategy of END not
needing to take up a whole value.  It makes chunk stack allocations
implicitly terminated by virtue of the pointer value indicating the
previous chunk.  More importantly than saving a REBVAL of memory
per chunk, it saves on the need to pay for that termination.