Disallow CONTINUE/WITH from an UNTIL
The definition of CONTINUE/WITH is that it goes to the top of loop
and acts as if the /WITH value had been what the body evaluated to.
Default for just a CONTINUE is to say the /WITH is UNSET!.

This definition throws a bit of a wrench into UNTIL, because UNTIL has
the same block as both body and condition.  It's clear that CONTINUE
inside of an UNTIL should continue the loop, but if it were treated as if
the body evaluated to UNSET! then that would be an error.

(In Rebol3 it is an error, Red currently returns from an UNTIL on unset.)

This modification maintains the status quo for Rebol3 by distinguishing
between CONTINUE (legal, keep looping) and returning an unset from
the "body-condition" (illegal, error).  It then disables the use of /WITH for
anything other than an UNSET!.  The rationale for this deviation is the
unique nature of UNTIL which merges both body and condition into the
same block.