BASE_CHUNK_SIZE as constant
For some reason that's not completely clear, the optimizer did not notice
that it could elide the ?: in the previous size calculation, and thus was
running a branch on something that it shouldn't have.  This dumbs it
down a bit so that the constant math can be seen more clearly and
cuts the time spent on the very-frequently-run-line in half.