Small Push_Chunk improvement, 25% faster
In the chunk stack implementation there were three branches that were
tested for, with one being the "null case" where no chunks were pushed.
In order to eliminate the need for testing that, this makes the initialize and
shutdown account for one empty chunk at the head.  Pushing a chunk to
the stack is about 25% faster as a result.

This also includes a minor speedup for realizing when a pointer is at the
head of a chunk, by keeping track of the current head vs. trying to
calculate the head from the chunk pointer.