Fix incorrect chunk checked for chunker
The chunk being dropped was being checked to get the new head
chunk, when the new head chunk should be the one for the top
chunk after the drop.

This also takes advantage of the new invariant that TG_Top_Chunk
is never NULL, due to the permanent empty chunk at the front of
the list.