Tweak frameless IF/UNLESS/EITHER to reuse D_OUT
If a cell other than the output cell is used to perform an evaluation into,
then that means you will have to pay the cost to proxy that cell into out
in the event of a throw.  In the case of IF, UNLESS, and EITHER there
is an additional cost because it is possible that a cell might not be a
block and hence not need to be evaluated.

Based on an observation that it's actually safe to do an array in a cell
into the same cell (if you aren't concerned about keeping the original
value after the evaluation), this changes the frameless implementations
of IF, UNLESS, and EITHER to use D_CELL as the scratch space in
the branch evaluations, as opposed to using D_OUT and then DOing
D_CELL into D_OUT.