Turn conditional falsehood into a OPT_VALUE bit
The test for conditional truth and falsehood required doing a type
check and hitting both the header and a secondary slot for the
logic.  This makes the two types that are conditionally false--NONE!
and LOGIC! with a value of FALSE!--carry a quick testable option
bit for the falsehood.  In the case of LOGIC! this option bit is also
the actual state of true or false, eliminating the need for anything
but a header initialization to SET_TRUE or SET_FALSE.