Delete stray unfinished OBJ_MODS feature
There was a stray routine with references to FLAGS_CLEAN which
does not exist, suggesting an attempt to track a dirty/clean bit on
object fields to be able to tell when they were modified.

Without investigating further to see how interesting such a feature is
or isn't, there simply isn't any significant code for its implementation
that couldn't be rewritten when the "modify with confidence" goals
have been met.

Deleted.