Rename REBSER.extra union to REBSER.misc
The term "extra" used in series routines has a fairly specific meaning,
which is the amount of additional capacity to add to an allocation during
a copy.  The "extra" field of a REBSER is a miscellaneous pointer-size
union that can hold series-specific data.

This field is about to play a more prominent role, by linking "frame"
series to their keylists...and also by serving as a terminator for compact
allocations (since pointers can serve as END markers, given they are
even on all platforms Rebol is known to run on).  So giving it a more
applicable name of "misc" helps avoid the confusion with the series
extra allocation space.