Cache type in parameter/arg macros in debug build
Rules for C require bitfields to be plain integers, and not an
enumerated type. This makes it inconvenient to see what the type
of a REBVAL is (one needs to cast it).

This caches the type in the debug build, coerced to the proper enum
type for viewing in the debugger.