Add series capture to userspace r3-legacy FOREACH
There is a feature in R3-Alpha, used by R3-GUI, which allows an
unusual syntax for capturing series positions (like a REPEAT or
FORALL) with a SET-WORD! in the loop words block of FOREACH:

    >> a: [1 2 3]
    >> foreach [s: i] a [print ["s:" mold s "i:" i]]

    s: [1 2 3] i: 1
    s: [2 3] i: 2
    s: [3] i: 3

This feature was removed from Ren-C due to it not deemed to be
"Quality", adding semantic questions and complexity to the C loop
implementation.  (e.g. `foreach [a:] [...] [print "infinite loop"]`)
That interferes with the goal of "modify with confidence" and
simplicity.

This implements the behavior in a userspace wrapper.  Should it
arise that MAP-EACH is used similarly in a legacy scenario then the
code could be factored and shared, but it is not likely that the
core construct will be supporting this in FOR-EACH or EVERY.

Longer-term, a rich LOOP dialect like Lisp's is planned:

    http://www.gigamonkeys.com/book/loop-for-black-belts.html