Fix single leaked "headroom chunker"
In converting the chunk stack to not have branches for the case of
an empty chunk stack (and having Init and Shutdown handle the fixed
case of one chunk always at the head), this introduced a possible
memory leak of one chunker.  Because it didn't go through the ordinary
freeing process that reclaimed headroom, just freeing the root chunk
didn't notice the headroom was still there and leaked it.