Make_Bitset for new series in shared op with binary
Because the implementation of a bitset is basically as a string of
binary bits (just like a binary), they share the implementation code
for performing bitwise operations.  However, they are still different
data types... and as one example of how they are different even today
a bitset can carry an "inverted" status bit.

However, the shared implementation was calling Make_Series to make
the series for the result.  This changes it to use Make_Bitset for bitsets
so that things like the inverted status bit can be initialized correctly.