Rough pass to eliminate FRAME! datatype
This is a running (but not well-factored) version of pushing through
the various logical steps required to get rid of the FRAME! datatype,
and instead put a value instance of whatever type the frame is
supposed to represent into the 0th cell of the frame.

It's a very basic baseline for further improvement, because a lot of
the mechanical consistency of the frames is being done by hand at
the callsites at the moment.  Assertions check for the correct patterns,
so the manual process is good while the proper abstraction layers
shake out with more things removed.