Alert more loudly on TRASH! access
The REB_xxx type enumeration value of 0 was taken away from
REB_END and given to REB_TRASH.  It is not meant to be a type
that is seen outside of debug builds.

This makes VAL_TYPE() assert in the debug build on seeing a trash
value.  It also makes it so that IS_TRASH is not built as one of the
standard bootstrap type checking macros, but rather made specially
as IS_TRASH_DEBUG--only available in debug builds.  SET_TRASH
is also renamed to SET_TRASH_IF_DEBUG...to make it more clear
that the operation is available in release builds but has no effect.

The noisier reaction to trash meant that the GC was tripping up on
checking a value it shouldn't have even really seen... the currently
being thrown argument.  So this moves that into a thread-local
variable.