ANY-OBJECT! => ANY-CONTEXT!
This change is motivated for many of the same reasons that the
ANY-BLOCK! => ANY-ARRAY! change was needed.  It's both a good
cognitive separation for users to have a distinct term for category vs. the
instance, and it makes it clearer in code to know when you have a thing
that isn't an OBJECT! but might be an ERROR! or PORT! or MODULE!
as well.

Synonym for the typeset is retained in mezz-legacy.