token-pasting to allow C keyword parameter names
The new PARAM/ARG/PAR and REFINE/REF macros generate
variable declarations for const structs that are compiled out of optimized
builds.  If the name fo the Rebol parameter or refinement collides with
a C keyword or other definition, then that creates a problem.

This uses a similar token pasting trick to what natives use such as
REBNATIVE(case) => N_case by prepending a "p_" in front of the
parameter.  This is only visible under the debugger.