Optional mixin argument to module
In order for refinement revocation to work, the refinement must be
UNSET!.  The statement:

    module/mixin hdr code (do-needs/no-user hdr)

Has a situation where do-needs may return NONE!.  This tries to
pass none as the argument for /mixin...which it cannot accept.  The
solution is to convert nones into optional via OPT:

    module/mixin hdr code (opt do-needs/no-user hdr)

That will correctly instruct the evaluator not to try to call module with
a /mixin argument, even though it is specified in the path.