Make Panic_Series debug only
Panic_Series is extremely useful, and it was conceived as a way
of having additional information during a panic that a debug build
could make use of while a release build would just panic normally.

In practice it hasn't yet been used in a very intentional way that fits
outside the scope of a debug build.  If it were to be used that way,
the panic would need a normal error as well with the series added.
At the moment, the scenarios in which this is useful are so few that
adding the parameter to Panic_Series seems like more of an
annoyance for the debug build...and a way to let calls placed for
debugging get left in where they don't belong.

So for now, this removes them from the debug build and cleans up
a couple of the aforementioned lingering references in the process.