REBSER, REBARR, REBMAP type incompatibility
(Though mostly searches and replace which are not intended to
change any functionality, this is still a very large commit.)

While Rebol leans heavily on the use of a core series type through
all the implementation, a very key distinction is whether a series
holds arbitrary binary bits -or- if it's a "Rebol Array" that holds
"Rebol Values". This distinction is of deep importance to basically
every client who has a series pointer in their hand, regarding
which operations to use with it.

In R3-Alpha the arrays were called "ANY-BLOCK!" and internally
the operations to work with them had the prefix BLK_ (BLK_HEAD,
VAL_BLK_SKIP, etc). This created confusion and several places
in the system lost track of whether they were working with an
operation that would apply equally well to parens or paths...and it
was hard to get one's bearings in the code as to what was a block!
and what wasn't. Moreover, there was no type incompatibility...so
it was left to runtime to find out when an array had been used as a
string or vice-versa.

Following on the heels of the separate REBFRM datatype, this is a
starting point of introducing a separate type-incompatible datatype
for arrays...which is the REBARR. It also adds one for maps, as
they are important to distinguish too. The nature of the change again
required touching basically every location that an array was used, and
some light cleanups are included which are not worth breaking into
separate commits.

This propagates the ARRAY name into the operations. Since there is
no ARRAY! concrete type in Rebol, it's easy to know that when one
is working with an array that it is either one of many array types (in
a REBVAL) or disconnected from any particular type at all (in a REBARR).

It also tries to help communicate when the index is being taken into
account... so instead of having an operation VAL_BLK_DATA(value)
which picks the vaulue at the embedded index for the ANY_ARRAY, the
operation is renamed to VAL_ARRAY_AT(value). Since AT is known to
require an index, the missing second parameter helps draw attention
to the fact that it's coming from the index in the value. (AT is preferable
to SKIP because it is shorter and works AT VAL_INDEX()...because in
C the indexes are 0-based.)

Additionally there is a start at trying to disentangle the use of the term
"tail", which alternately used to refer to the length as well as the memory
address of the value one-past-the-end of the valid content of a series.
LEN is adopted as the preferable term for the length, while the pointer
interpretation is propagated as a complement to HEAD.

The organization of series routines into a separate system header and a
better documentation / layout is pending.