Remove unused At_Head, %sys-panics.h
The At_Head function is unused, and claims it is "useful for functions
that need a block at its head", by making a copy if it is not at the head
and returning the series as-is otherwise.  This is actually *not* useful
because the semantics of writing a routine that doesn't know if it's
making a copy or not are bad.  Routines generally do one or the other,
and it also messes with the managed/unmanaged expectations of the
resulting series.

Also removes the now-empty %sys-panics.h file, as the panic()
mechanism has been unified with regular fail() errors.