Move IS_SET and IS_SCALAR into %make-boot.r
IS_SET and IS_SCALAR aren't checking for exact type values,
but using comparison operators of the derived values.

The macros and checks related to the relative numbering of types
need to be centralized together, so that if types are renumbered
the checks can be looked at all in one place...as the bugs will
not necessarily be obvious.