Get rid of obsolete STR_ routines
STR_XXX appears to be a holdover synonym for BIN_XXX that existed
prior to the unicode division.  Strings having a variance between byte-size
and REBUNI-size elements is implemented above the series level, so there
is no STR_XXX series operation that is adaptive to the size.  (SERIES
operations in general adapt to the width, but don't know the type to return.)

This removes the routines, and in cases that weren't obviously known to be
binary series adds an assert that the input is BYTE_SIZE.  This way, if those
routines were getting an input that was either REBUNI width or REBYTE
width, there will be a notification...instead of giving a corrupt offset (which is
what would have been happening before).