Disambiguate VAL_SET and SET_TYPE
When working early on with the Rebol source, one of the confusing things
to learn at first was the difference between VAL_SET and SET_TYPE.
Both took a REB_XXX type.  Between the two, VAL_SET was the more
frequently used...and also the more confusing...as VAL_SET is very
close to SET_VAR.  So it's hard to get one's bearings just by looking at
it, even if one has some grasp of the basics.

Over time the difference sunk in...that VAL_SET resets the entire header
(not just the type), clearing all the value options flags and other bits.  Then
SET_TYPE left all the option bits alone and modified only the type.

Despite that knowledge, the clearing of the all the header flags is an
important distinction...and sometimes problems that were hard to find came
from realizing a flag had been cleared.  Rather than favor brevity, Ren-C
favors clarity so that newcomers to the source have a better shot at getting
their bearings, so these are renamed:

    VAL_SET => VAL_RESET_HEADER
    SET_TYPE => VAL_SET_TYPE

There is a tradeoff here as to whether it should be prefixed with VAL_ or
be SET_VAL_TYPE and RESET_VAL_HEADER, which ties into the
concern which makes that seem more like one is doing a SET_VAR type
operation as opposed to a Val_Init type operation.  This may be reviewed in
the future, but it's clear enough for now.