Eliminate SET_INT32 macro
The SET_INT32 macro is confusing because unlike SET_INTEGER, it
does not set the type of the REBVAL it is given to REB_INTEGER.  It only
sets the 64-bit value to a 32-bit number.

It seems to have come into existence because VAL_INT32--due to a cast--
produced an Rvalue instead of an assignable LValue.  So:

    VAL_INT64(v) = some_64_bit_int; // works
    VAL_INT32(v) = some_32_bit_int; // doesn't work

It could be fixed by making the macro do VAL_RESET_HEADER() to a
REB_INTEGER, but the macro is used only 3 times and isn't really that
compelling, given that Rebol takes for granted the existence of 64-bit
integer assignments.  Not C89 standard, but long long is available in most
compilers...but this macro wasn't useful for any purpose of bridging gaps
to compilers without long long.  So this deletes it and just uses an assign
of a 32-bit source to a 64-bit target as is expected to work (and was
expected to work by the macro).