Rename VAL_TAIL => VAL_LEN_HEAD, make rvalue
In R3-Alpha, there were two operations for getting the length of an
ANY-SERIES! REBVAL.  One was VAL_TAIL and the other was VAL_LEN.

The difference was that VAL_LEN would take into account the index
position.  So if you had the series [a b c d] at index 2 (in C, zero based)
then the VAL_LEN would be 2, for [c d].  However VAL_TAIL would
give back the entire length of the series, so 4.

This distinction was all the more confusing because TAIL was used to
give back a pointer to the terminator in other contexts.

VAL_LEN has already been renamed to VAL_LEN_AT, in order to stress
that there is an index position taken into account.  Since there is only one
parameter (the ANY-SERIES! value) then the "at" index can only be coming
from one place...the value.  This changes VAL_TAIL to be VAL_LEN_HEAD
because its answer is what you would get from `length head series` if you
were in the interpreter.

Because the series abstraction is about to get a bit more complex, this adds
in the twist of not being able to assign VAL_LEN_HEAD directly.  Rather, one
has to go through the SET_SERIES_LEN macro.  This becomes important
in a later commit.