Access series ->tail and ->data through macros
The Rebol series REBSER was designed in such a way that it
would do things like concretely manipulate its data pointer when
it was "biasing" it forward.  A client could really just accept that
the ->tail was the length and the ->data pointed to the first
element.  So most would get and set these pointers directly...not
using any macros or abstractions to do it.

In Ren-C the desire is to have a very light abstraction which permits
the data pointer for short series to be resident in the series node
itself.  This means getting the data pointer and length needs to be
done as a check for a bit before deciding to reach into the dynamic
portion of the content for the pointer.

Having SET_SERIES_LEN go through a macro offers another benefit,
which is that the debug build can make sure that the length is not
being set to something beyond the known capacity of the series.