Call REBVAL components "header"/"payload"
Previously the two components of a REBVAL were called "flags"
and "data".  The header contains a type (which is an enumerated
value and not a flag), and data is also a very generic term most often
used for data pointers (as in series->data).  The terms "header" and
"payload" are a better fit, with the header containing "bitfields"... some
of which happen to be flags.

(This is part of an impending wave of expository documentation, and
it helps if the terms used in the writeup match the terms for the structs
and field names.)