Assert no odd size in series data
The BYTE_SIZE() routine for testing if a series has a width of 1 uses
a test of only the low bit of the width.  A comment on BYTE_SIZE() says
it can do this because no other odd values are supported as sizes.

Whether that is a good or bad tradeoff to be making aside, there was no
assert checking it...so someone making a series with odd sized elements
besides 1 would get a surprise when it came back reporting it was
BYTE_SIZE().  This adds the assert.

Includes a few formatting changes to m-pools.c.