Groundwork for multiple series node types
A series node (REBSER) is the fixed-size portion of a series, which
is allocated from a memory pool of REBSER-sized items.  During the
lifetime of a series, its data array may have to be freed and reallocated
many times based on varying needs for capacity--likely getting a new
address each time.  However, the pointer to the small REBSER (a "Node"
from the memory pool, in Rebol and Red terminology) will remain constant
for the lifetime of the series, until all references that use that pointer have
been released and it is garbage collected.

Historically there was a "one-size-fits-all" REBSER structure, which had
enough information to track a dynamic allocation of memory.  So it had
the pointer, a total allocated capacity, the length in use, the amount that
the pointer had been "biased" due to removals of data at the head (which
created a "head capacity" or bias).

This commit pushes forward on the movement of REBSER to a new file
specifically for series definitions (%sys-series.h), and also on the idea of
using a union to have multiple different series node types.  One kind of
series is a kind where the data itself fits entirely in the node.  Other future
categories would be those that represent copy-on-write "clones", which
are also intended to be used as a replacement for today's symbol table.

The dynamic variant is still the only one implemented, but this lays the
groundwork for being able to work on the other REBSER types.  It also
renames "tail" to LEN, consistent with pushes to be more clear on the
difference between those two words.  (TAIL as pointer/position, LEN as
integer number.)