%sys-value.h intro text, forward defs, light renaming
This commit includes a start on an introduction to what a REBVAL
is, and also does a few renamings:

    Reb_Position => Reb_Any_Series
    Reb_Context => Reb_Any_Context
    Reb_Word => Reb_Any_Word
    Reb_Function.func => Reb_Any_Function.impl

The "Any" helps disambiguate Reb_Series (the REBSER type) from
the value payload type.  Making the naming system consistent with the
value categories helps emphasize that the structure is used for all values
of that type, e.g. there's not a separate Reb_Set_Word struct.