Clean up superfluous REBNATIVE comments
In the great source conversion of 2015, the native specs found a new
home in the comments of the C file defining the natives.  This made them
easier to edit alongside the code without having to go find them, and also
served a documentary purpose.

A few REBNATIVE() definitions had a little bit of commentary that could be
redundant with the spec, or perhaps add information that should be in the
spec.  This tidies up those few definitions.