Add header documentation, change type to 6 bit
In the bitflags, 8 bits were being used for the type when really there
can only be 64 low-level types--as long as typesets are bitsets of 64
bits, and all those types need to be able to fit in a typeset.

There may be some performance advantage to masking out a power
of 2 bits, although it's not exactly clear what that would be.  If that
is shown in timings then perhaps it's worth losing the extra 2 bits, but
if not the bits are good to reclaim.

One potential use marked out is for the "lit-bit", which would indicate
if a type was literal or not... allowing literal integers/parens/blocks/etc.