Document #pragma pack, use push/pop
The use of `#pragma pack(4)` in %reb-value.h is probably not needed,
and if it is then it should be explained better why.  This starts the
process with some notes, and also uess the push/pop convention so
that if code came in with a different packing pragma then it will be
maintained after including %sys-value.h, instead of reset.