Add precursor of check for "formatted" values
This is the beginning of functionality that will make debug builds
able to check to make sure they have the right to write over an END
marker with a REBVAL (or even another end marker).  This will make
it safer to use the trick of having pointers internal to REBVAL array
containers do double duty as end markers for the contained data when
it's at the full capacity.

It also addresses an issue with the C++ build and use of integers in
enum slots.  However it does include the use of REB_TRASH as a
value in release builds, which needs to be addressed another way.