Make PG_Word_Table.array start with typeset, not none
In experimenting with NONE!'s value payload holding some debug
information, it was discovered that there is a usage of NONE in the
0 slot of the word table which appears to be invalid...and if the none
has something other than 0 in its payload there is a problem.  What
was happening previously was that a CLEAR operation on an array
of data ensured that the NONE! payload (which otherwise had no
guarantee historically) would have a 0 in the symbol position...so
Collect_Keys would later read that zero symbol out.

This formalizes the process by using an actual typeset with a SYM_0
symbol key instead of a none...allowing NONE! to make what use of
its payload it wishes.