TRASH!, UNSET!, NONE!, LOGIC! notes/updates
This provides commentary and some new features for the types
that have no payloads.  Whereas before only TRASH! would save
the file and line number in the payload of where it was assigned, now
all 4 types that don't have need for their payload do it.  This will be
very useful for UNSET!, maybe useful for NONE! and LOGIC! as
well under some circumstances.

The common tracking payload (Reb_Track) is also extended to save
the do "tick count" when the value was created.  This can help
pinpoint exactly where an unexpected trash/unset/none/logic was
made, so that a breakpoint can be set in a repro case on that very
instruction.