Some reorganization of %sys-value.h
This is not at a final state--and in fact some areas and comments
have been left "messy" to help serve as a reminder that those sections
still need going over.  However, it does put things in slightly more
order, with core concepts at the top...and things that will likely be
excluded from Ren-C library builds (like GOB!) nearer to the bottom.