Fix bug in Find_Key not dividing length by "wide"
In what appears to be a longstanding bug in R3-Alpha, the method
for adding indices into maps did not take into account the "width".
While when decoding indices it would subtract one and multiply by
the wide, the division was missing from the addition.  This caused
accesses into invalid memory, and likely was the cause of many bugs.