Make parse INTO recognize END
New rules prohibit the testing for a value that might be an END marker
for its type before it has been vetted for not being an END.  The parse
INTO operation was previously failing on end values by virtue of them
not being blocks or strings, and not testing for end first.  This adds the
needed test.