Enhance Assert_Not_In_Series_Data, lock closures
This improves the behavior of Assert_Not_In_Series_Data to take
locked series into account, and locks closures while they are on
the stack.  This prevents false positives in the (heavy) check on if
a series slot is passed as the output location for an evaluation, which
could spell trouble as that location could be on the stack arbitrarily
long, during which time a series may be expanded or contracted.

(Or a future garbage collector might just do background cleanup if
it believes a series's data is not currently referenced by code on
the stack.)

This improvement removes the need to use a series data size finder,
which has been given a more accurate name and usage (and the
GC_Stay_Dirty bit killed in the process.)