Convert PARAMLIST series into framelike FUNC
In the evolution of the solution to definitional return and eliminating
the FRAME! datatype, function parameter lists were the first series
that switched to include the ANY-FUNCTION! value in the [0]th
element, in order to reconstitute the function value from a pointer
in definitional return.

Given that they now have this flexibility, it makes sense to allow an
ANY-FUNCTION! to be passed around by pointer instead of by value
with the same convenience that a REBFRM can be passed around.
This begins that transition, retaking the name REBFUN (formerly
used for natives, now called REBNAT).

!!! The FFI currently has some kind of garbage collection bug, which
is being investigated.