Fix to image! gob
Before this change, it tries to rendered the top window gob containing
the gob in question, and clip the image by the area of the gob, which
will contains everything in the whole gob tree that falls into the area.

This commit changes it to only render the gob and its descendants.