single-pointer FUNC in Do_Core instead of REBVAL
This leverages the FRAME-like "FUNC" that knows the canon identity
of its value in Do_Core, as a test of the viability of the concept.

An observed performance regression indicates that the change itself
to Do_Core may need to be backed out to go back to caching a full
REBVAL.  Regardless of the decision whether to do that (relatively minor)
change back or not, the supporting test changes for the feature will
be kept.