Switch header to use standard bitmasking, native size
The original definition of the header for Rebol values predated the
64-bit build, and did something that is undefined in C.  It used a switch
based on the endianness of the processor to decode an unsigned 32-bit
integer as bitfields in either increasing or decreasing bytes.  As it happens,
the endianness of the processor is not the only issue which could affect
the bit fields...and also you can't read from the integer in the union if the
bit fields have been assigned.

Also, by using a 32-bit integer in what would be a 64-bit slot on 64-bit
architectures, it slows down header reads and writes in order to be able
to break out half of the information from the fetch.  This switches to use
Rebol's equivalent of an intptr_t and do ordinary masking, which
dispenses with the issues altogether.

The 6-bit datatype fits particularly well with the 2 low bit settings for the
end flags, so those constitute the lowest byte.