Object comparison ignores hidden fields
Pursuant to being able to have a "real" definitional SELF (vs a
hardcoded hack on the SYM_SELF in bind), handling of hidden
fields in objects needs to be beefed up.  This allows two objects
which have hidden fields to compare equal if their non-hidden
fields line up.

This means object comparisons cannot short-circuit based on
a count of the number of fields.  (Unequal objects should
in the average case come across as unequal quickly enough just
based on keys or values not matching...and to the extent that
they would not, they would perhaps be pathologically just as
likely to be equal lengths...moreso probably.)