Simplify FRAME creation during bootstrap
The abstraction of a frame is some bookkeeping on top of two series
that are maintained in parallel to represent variable definitions (keys)
and variable values (vars).  The basic construction of frames involves
using internal structures like the collect buffer, which must be allocated
and configured in order to work.  So there is a bootstrap problem in
creating the frame that represents the Root and Task contexts, because
they can't call the higher level frame creation routines.

Historically this meant that they built frames by hand out of series in a
two phase process: the varlist created first, with the keylists created
later.  The manual process wound up repeating several elements of
frame creation, which had to be kept in sync and made the code harder
to follow.

This makes it a bit easier by using the basic frame creator Alloc_Frame,
and then freeing the keylist.  It later uses the high-level routine for making
keylists to put it back with the necessary data, once the service is ready.
The code redundancy is cut down and the bootstrap process does not
have to worry about what the manual process of building frames out of
series entails.