Collect_Keys fails safely (reset binds, BUF_COLLECT)
Because the binding table and collect buffer would be left in a bad
state, the use of a "trap" (longjmp) was discouraged in the comments
of Collect_Keys.  Yet two spots did fail, and had to handle the cleanup
manually.

Beyond the inconvenience of such manual work every time you need
to raise an error, It's not really viable to say that you "can't fail".  The code
is doing memory allocation (which expanding the collect buffer does) and
any allocation might fail.

This is a quick and dirty addition of resetting the bind table to all 0 and
cleaning out the collect buffer in the trap handler, and removing the manual
code from the existing failure cases.