ASSERT_FRAME notices embedded frame mismatch
A frame's "rootvar" (value in the [0] index) is not a variable in the
frame corresponding to a key symbol.  Rather it is a copy of the
REBVAL of whatever ANY-CONTEXT! it is a frame for, allowing the
value (sizeof(void*) x 4) to be reconstituted from the frame pointer
alone (sizeof(void*)).

One of the elements of that REBVAL is a pointer to the frame itself.
This adds to ASSERT_FRAME a check that when asserting a frame
that the embedded pointer is the same as the frame pointer that is
currently being checked.